+++ tomcat/src/main/java/org/apache/coyote/http11/constant/HttpCookie.java
@@ -14,22 +14,11 @@ public class HttpCookie {
         }
     }
 
-    public boolean containsJSessionId() {
-        return cookies.containsKey("JSESSIONID");
-    }
-
-    public String getJSessionId() {
-        if (!cookies.containsKey("JSESSIONID")) {
-            throw new IllegalStateException("JSESSIONID가 존재하지 않습니다.");
-        }
-        return cookies.get("JSESSIONID");
+    public boolean contains(String key) {
+        return cookies.containsKey(key);
     }
 
     public String get(String key) {
         return cookies.get(key);
-    }
-
-    public String getCookieString() {
-        return null;
     }
 }
